gcc*-libcxx: move subports to a sperate Portfile
The build mechanism for gcc*-libcxx is very different from
GCC and Libgcc, so there have be a fair number of `if` statements
to keep them in the same Portfile.

The version numbers of gcc*-libcxx do not need to be correlated
with the version numbmers of GCC and Libgcc.